fix: 添加部门删除前检查子组织逻辑

+++ server/src/main/kotlin/math/yl/love/database/service/DepartmentService.kt
@@ -59,6 +59,10 @@ class DepartmentService(
 
     @Transactional(rollbackFor = [Exception::class])
     override fun removeById(id: Serializable): Boolean {
+        val child = queryWrapper.eq(Department::parentId, id).list()
+        if (child.isNotEmpty()) {
+            throw BizException("该组织下有子组织，无法删除!!!")
+        }
         return super.removeById(id) && userDepartmentService.removeByDepartmentId(id)
     }
 }
